added the comparing runtime of 20 Bits and 40 Bits

+++ src/PrimeChecker.java
@@ -1,5 +1,5 @@
 public class PrimeChecker {
-    public static boolean isPrime(int n) {
+    public static boolean isPrime(long n) {
         if (n <= 1) {
             return false;
         }
